backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai>=1.30.0
yt-dlp>=2024.4.0
fpdf2>=2.7.8
python-dotenv>=1.0.1
